+++ src/main/java/com/allthemods/gravitas2/GregitasGTAddon.java
@@ -1,5 +1,6 @@
 package com.allthemods.gravitas2;
 
+import com.allthemods.gravitas2.data.recipe.GregitasRecipes;
 import com.allthemods.gravitas2.data.tag.GregitasTagPrefixes;
 import com.allthemods.gravitas2.material.GregitasMaterials;
 import com.allthemods.gravitas2.recipe.capability.GregitasRecipeCapabilities;
@@ -9,6 +10,9 @@ import com.gregtechceu.gtceu.api.addon.events.KJSRecipeKeyEvent;
 import com.gregtechceu.gtceu.integration.kjs.recipe.components.ContentJS;
 import com.mojang.datafixers.util.Pair;
 import dev.latvian.mods.kubejs.recipe.component.NumberComponent;
+import net.minecraft.data.recipes.FinishedRecipe;
+
+import java.util.function.Consumer;
 
 @GTAddon
 public class GregitasGTAddon implements IGTAddon {
@@ -30,6 +34,11 @@ public class GregitasGTAddon implements IGTAddon {
     @Override
     public void registerMaterials() {
         GregitasMaterials.init();
+    }
+
+    @Override
+    public void addRecipes(Consumer<FinishedRecipe> provider) {
+        GregitasRecipes.addRecipes(provider);
     }
 
     public static final ContentJS<Double> PRESSURE_IN = new ContentJS<>(NumberComponent.ANY_DOUBLE, GregitasRecipeCapabilities.PRESSURE, false);