+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - buff EBF heating, disable ingot cooling with water
 - actually fix EBF heat number
+- make GT ore blocks able to collapse